requirements.txt
apify_client==1.10.0
langchain==0.3.25
langchain_google_genai==2.1.5
langgraph==0.4.8
pydantic==2.11.5
python-dotenv==1.1.0
streamlit==1.45.1
